+++ isitfit/cost/ec2/pipeline_factory.py
@@ -12,6 +12,8 @@ def ec2_cost_analyze(ctx, filter_tags, save_details):
     from isitfit.cost.ec2.ec2Common import Ec2Common
     from isitfit.cost.redshift.iterator import Ec2Iterator
 
+    from isitfit.utils import TqdmMan
+    tqdmman = TqdmMan(ctx)
 
     share_email = ctx.obj.get('share_email', None)
     ul = CalculatorAnalyzeEc2(ctx, save_details)
@@ -26,15 +28,18 @@ def ec2_cost_analyze(ctx, filter_tags, save_details):
     mm = MainManager(ctx)
     ec2_cat = Ec2Catalog()
     ec2_common = Ec2Common()
-    ec2_it = Ec2Iterator(filter_region=ctx.obj['filter_region'])
+    ec2_it = Ec2Iterator(ctx.obj['filter_region'], tqdmman)
 
     # boto3 cloudtrail data
-    cloudtrail_manager = CloudtrailCached(mm.EndTime, cache_man)
+    cloudtrail_manager = CloudtrailCached(mm.EndTime, cache_man, tqdmman)
 
     # update dict and return it
     # https://stackoverflow.com/a/1453013/4126114
     inject_email_in_context = lambda context_all: dict({'emailTo': share_email}, **context_all)
     inject_analyzer = lambda context_all: dict({'analyzer': ul}, **context_all)
+    def inject_tqdmClose(context_all):
+      mm.gtqdm.close()
+      return context_all
 
     # utilization listeners
     mm.set_iterator(ec2_it)
@@ -52,6 +57,7 @@ def ec2_cost_analyze(ctx, filter_tags, save_details):
     mm.add_listener('all', ul.after_all)
     mm.add_listener('all', inject_analyzer)
     mm.add_listener('all', ra.postprocess)
+    mm.add_listener('all', inject_tqdmClose)
     mm.add_listener('all', ra.display)
     mm.add_listener('all', inject_email_in_context)
     mm.add_listener('all', ra.email)
@@ -72,6 +78,8 @@ def ec2_cost_optimize(ctx, n, filter_tags):
     from isitfit.ec2_catalog import Ec2Catalog
     from isitfit.cost.ec2.ec2Common import Ec2Common
     from isitfit.cost.redshift.iterator import Ec2Iterator
+    from isitfit.utils import TqdmMan
+    tqdmman = TqdmMan(ctx)
 
     ol = CalculatorOptimizeEc2(n)
 
@@ -85,14 +93,17 @@ def ec2_cost_optimize(ctx, n, filter_tags):
     mm = MainManager(ctx)
     ec2_cat = Ec2Catalog()
     ec2_common = Ec2Common()
-    ec2_it = Ec2Iterator(filter_region=ctx.obj['filter_region'])
+    ec2_it = Ec2Iterator(ctx.obj['filter_region'], tqdmman)
 
     # boto3 cloudtrail data
-    cloudtrail_manager = CloudtrailCached(mm.EndTime, cache_man)
+    cloudtrail_manager = CloudtrailCached(mm.EndTime, cache_man, tqdmman)
 
     # update dict and return it
     # https://stackoverflow.com/a/1453013/4126114
     inject_analyzer = lambda context_all: dict({'analyzer': ol}, **context_all)
+    def inject_tqdmClose(context_all):
+      mm.gtqdm.close()
+      return context_all
 
     # utilization listeners
     mm.set_iterator(ec2_it)
@@ -109,6 +120,7 @@ def ec2_cost_optimize(ctx, n, filter_tags):
     mm.add_listener('all', ec2_common.after_all)
     mm.add_listener('all', inject_analyzer)
     mm.add_listener('all', ra.postprocess)
+    mm.add_listener('all', inject_tqdmClose)
     mm.add_listener('all', ra.display)
 
     return mm