+++ isitfit/cost/pull_cloudtrail_lookupEvents.py
@@ -13,7 +13,7 @@
 import datetime as dt
 from dateutil.relativedelta import relativedelta
 import boto3
-from tqdm import tqdm
+#from tqdm import tqdm
 import json
 
 import logging